# public/doc/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Supp",
        "description": "An API that uses supp service",
        "version": "1.0.0"
    },
    "basePath": "/api",
    "schemes": [
        "http"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/tokens": {
            "put": {
                "tags": [
                    "token"
                ],
                "summary": "Refresh token for authorization",
                "description": "",
                "operationId": "updateToken",
                "parameters": [
                    {
                        "name": "token",
                        "in": "formData",
                        "description": "working token",
                        "required": true,
                        "type": "string",
                        "maxLength": 255
                    }
                ],
                "responses": {
                    "202": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "token": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Bad: Token not provided",
                        "schema": {
                            "$ref": "#/definitions/ErrorModel"
                        }
                    },
                    "403": {
                        "description": "Forbidden: invalid token provided",
                        "schema": {
                            "$ref": "#/definitions/ErrorModel"
                        }
                    },
                    "404": {
                        "$ref": "#/responses/NotFoundResponse"
                    },
                    "401": {
                        "$ref": "#/responses/NotAuthResponse"
                    },
                    "default": {
                        "$ref": "#/responses/DefaultErrorResponse"
                    }
                }
            },
            "post": {
                "tags": [
                    "token"
                ],
                "summary": "Create token for authorization",
                "description": "",
                "operationId": "createToken",
                "parameters": [
                    {
                        "name": "email",
                        "in": "formData",
                        "description": "email",
                        "required": true,
                        "type": "string",
                        "maxLength": 255
                    },
                    {
                        "name": "password",
                        "in": "formData",
                        "description": "password",
                        "required": true,
                        "type": "string",
                        "maxLength": 255
                    }
                ],
                "responses": {
                    "201": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "token": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "404": {
                        "$ref": "#/responses/NotFoundResponse"
                    },
                    "401": {
                        "$ref": "#/responses/NotAuthResponse"
                    },
                    "default": {
                        "$ref": "#/responses/DefaultErrorResponse"
                    }
                }
            }
        },
...
    },
    "definitions": {
        "ErrorModel": {
            "required": [
                "status_code",
                "message"
            ],
            "properties": {
                "status_code": {
                    "type": "integer",
                    "format": "int32"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
    "responses": {
        "NotAuthResponse": {
            "description": "Not Auth",
            "schema": {
                "$ref": "#/definitions/ErrorModel"
            }
        },
        "NotFoundResponse": {
            "description": "Not Found",
            "schema": {
                "$ref": "#/definitions/ErrorModel"
            }
        },
        "DefaultErrorResponse": {
            "description": "Default Error",
            "schema": {
                "$ref": "#/definitions/ErrorModel"
            }
        }
    },
    "securityDefinitions": {
        "token": {
            "type": "apiKey",
            "name": "token",
            "in": "query"
        }
    }
}